agi/beta/requirements.txt
# Dependencies for the AGI project
openai>=1.0.0
python-dotenv>=1.0.0